Add a set of binary tree functions
These functions are just a thin wrapper around Glibc's TSEARCH(3) binary
tree functions.

This introduces a new type, ac_btree_t which is an opaque structure. You
would create a new binary tree like

    ac_btree_t *tree;

    tree = ac_btree_new(compare, free_tnode);

where compare is a pointer to a function that is used to compare nodes
in the tree and free_tnode is a pointer to a function to free the node
memory, this can be specified as NULL if no memory needs to be free'd
etc. The other functions are

    ac_btree_add()
    ac_btree_lookup()
    ac_btree_foreach()
    ac_btree_remove()
    ac_btree_destroy()

See the btree_test() function in test.c for examples of usage of these
functions.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>