Add a couple of network name lookup functions
Add ac_net_ns_lookup_by_host() and ac_net_ns_lookup_by_ip() functions.
These take a pointer to a struct addrinfo that will be passed to
getaddrinfo(3)/getnameinfo(3), a pointer to the string containing the
item to be looked up and a callback function that is called for each
returned item.

See the net_test() function in test.c for example usage.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>