Add a set of circular queue functions
This adds a simple implementation of a circular queue with the following
API

    ac_cqueue_new()
    ac_cqueue_push()
    ac_cqueue_pop()
    ac_cqueue_foreach()
    ac_cqueue_is_empty()
    ac_cqueue_nr_items()
    ac_cqueue_destroy()

This also introduces a new type, ac_cqueue_t, which is an opaque
structure that is meant to be accessed only via the above functions.

A new circular queue would be created like

     ac_cqueue_t *queue;

     queue = ac_cqueue_new(SIZE, FREE_FUNC);

Where SIZE is the number of items the queue can contain and FREE_FUNC is
a pointer to a function that is used to free the memory of an item. This
can be NULL if no freeing is needed.

The queue is implemented such that it won't overwrite existing items if
the queue is full, rather ac_cqueue_push() will just return -1.

See the cqueue_test() function in test.c for example usage.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>