ac_cqueue: Change a couple of else if's into else's
When iterating over the circular queue there are just two cases we need
to cater for; cqueue->front < cqueue->rear and also
cqueue->front >= cqueue->rear. The first one is in an if () so no need
to spell the second one out in an else if (), but just use else.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>