Add a set of quark (string to integer mapping) functions
This is inspired by the GQuark of glib. It maps strings to unique
integers. It adds the following four functions

    ac_quark_init()
    ac_quark_from_string()
    ac_quark_to_string()
    ac_quark_destroy()

It also introduces a new type, ac_quark_t, which is an opaque structure
for holding the quark data.

A new quark would be created like

    ac_quark_t quark;

    ac_quark_init(&quark, NULL);

The second parameter is an optional pointer to a function to free the
quark itself when ac_quark_destroy() is called and is not usually
needed but can be used like

    ac_quark_t *quark;

    quark = malloc(sizeof(ac_quark_t));
    ac_quark_init(quark, free);

The quark itself is built on top of an ac_btree_t.

Quark id's start from 0.

See the quark_test() function in test.c for example usage.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>