ac_misc: Make ac_misc_ppb() thread safe
ac_misc_ppb() was not thread safe due to the use of global (albeit
constrained to ac_misc) variables for storing the bytes factor and
prefix.

Re-work it to get rid of those completely and create a set of macros
and a new internal function to handle the setting of the factor and
prefix text.

This simplifies the code providing a nice code reduction.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>