Add a set of singly linked list functions
This is somewhat based on the GSList from glib. It adds the following
public functions

    ac_slist_add()
    ac_slist_preadd()
    ac_slist_remove()
    ac_slist_reverse()
    ac_slist_find()
    ac_slist_foreach()
    ac_slist_destroy()

It also introduces a new type, ac_slist_t, which is an opaque structure
for working with singly linked lists.

There is no function to initialise a new singly linked list, you simply
create a new variable of type ac_slist_t and set it to NULL, e.g

    ac_slist_t *mylist = NULL;

See the slist_test() function in test.c for example usage.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>