ac_fs: Add a file copy function
Add ac_fs_copy(), that takes three arguments; the file to copy, the
destination and flags which can either be 0 or AC_FS_COPY_OVERWRITE to
allow the destination to be overwritten.

This is implemented using sendfile(2) and should work fine if your using
a 2.4 or >= 2.6.33 kernel.

See the fs_test() function in test.c for example usage.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>