Add a couple of geospatial related functions
This adds the following functions

    ac_geo_haversine()
    ac_geo_vincenty_direct()

This also introduces some new types. The first is ac_geo_ellipsoid_t
which is an enum of ellipsoid reference shapes, e.g WGS-84 and ac_geo_t
which is a structure to hold lat/lon/bearing etc.

The first function uses the Haversine formula to calculate the distance
between two points on Earth.

The second function uses Vincenty's direct formula to calculate the end
point and bearing given a starting point, bearing and distance.

See the geo_test() function in test.c for example usage.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>