ac_geo: Add functions to convert between decimal degrees and DMS
This adds a couple of functions to convert to and from decimal degrees,
e.g 38.8897 and DMS (degrees, minutes & seconds) e.g 38° 53′ 22.92″.

The two functions added are

    ac_geo_dd_to_dms()
    ac_geo_dms_to_dd(()

A new type is also introduced, ac_geo_dms_t which is a structure to
hold the broken out degrees (int), minutes (int) & seconds (double).

Note: It is up to the user to track if the decimal degrees being
converted is negative or if the DMS is North/South of the Equator or
East/West of the prime meridian.

See the geo_test() function in test.c for example usage.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>