Change the soname to libac.so.0
Previously the soname was libac.so.MAJOR.MINOR however this meant that
each time the library MINOR version was updated you needed to recompile
programs that used it even even though the changes were backwards
compatible.

So make the soname stop at the MAJOR version. Yes it's early days
version 0. so breakage may be expected anyway (though there hasn't
been).

In the rpm spec file we don't need to create the soname symlink
(libac.so.MAJOR) as that is created by ldconfig. We just need to create
the linker name (libac.so) symlink.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>