ac_fs: Use fallocate(2) in ac_fs_copy()
fallocate(2) is used to tell the file system to pre allocate a certain
amount of space. This *may* improve performance when writing the file,
it will also ensure there is enough space left in the file system and
also that the file is laid out contiguously.

If fallocate is not supported by the underlying file system then we fall
back to ftruncate(2) instead.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>