ac_fs: Keep right errno from ac_fs_copy()
Commit a8f7ba9 ("ac_fs: Fix a potential fd leak in ac_fs_copy()") caused
a slight issue in that if we failed to open the output file, we would
then exit the function and along the way try to close this file using
-1 for the file descriptor meaning errno would be set to EBADF from
close(2) rather then the errno from the failed open(2).

After this fix we should now error out with the correct errno for the
bit that failed.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>