Compile with -Wextra
Add -Wextra to the CFLAGS. This should ensure the code is kept a little
tighter, which is no bad thing for a library.

For the test program we disable the -Wunused-parameter check as that
triggers a few times (intentionally) and for this case it seemed cleaner
to do that than sprinkle __attribute__ ((unused))'s all over the place.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>