ac_misc: Add a wrapper around crypt_r(3)
This adds ac_misc_passcrypt() which can be used to produce hashed
passwords using the crypt(3) function. We use the re-entrant variant
crypt_r for thread safety.

This introduces a new type, ac_hash_algo_t which is an enum defining
various hashing algorithms, currently MD5, SHA256 & SHA512 which can be
used by crypt() and also SHA1.

This also introduces a typedef for struct crypt_data, ac_crypt_data_t,
which is used by crypt_r() for storing state (and the hashed password)
of which the initialized element needs to be set to 0 upon the *first*
call to crypt_r(). So to use this function do something like

    ac_crypt_data_t data = { .initialized = 0 };
    char *crypted;

    crypted = ac_misc_passcrypt("asdfghk", AC_HASH_ALGO_SHA256, &data);
    /* do something with crypted */
    crypted = ac_misc_passcrypt("kjhgfdsa", AC_HASH_ALGO_SHA256, &data);
    /* do something with crypted */

This will create hashed passwords for 'asdfghk' & 'kjhgfdsa' using the
SHA-256 algorithm.

NOTE: The pointer returned actually points into the 'data' structure.
This does *not* need to be free'd.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>