Add some valgrind suppression rules
I try to keep libac valgrind *clean*, however with the previous commit
e47a428 ("ac_misc: Add a wrapper around crypt_r(3)") valgrind started
complaining about un-reachable memory coming from dlopen(3). This seems
to be a somewhat common occurrence and it seems it can generally be
ignored.

So we add a couple of valgrind suppression rules and a .valgrindrc file
in the src/ directory that points valgrind to this suppression file, so
when valgrind is run within the src directory it will automatically pick
these up.

You could also add a pointer to these in ~/.valgrindrc to pick these up
when using libac in other projects.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>