ac_fs: Check for unknown flags in ac_fs_copy()
Add a check for unknown flags passed to ac_fs_copy(), this can be useful
for applications to check if a certain flag is supported by the version
of library currently in use (as an alternative to using the library
version number).

If we detect any unknown flags we set errno to EINVAL and return -1.

Also see https://lwn.net/Articles/588444/ for a more detailed rationale.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>