Compile with -fvisibility=hidden
This causes symbols to be 'hidden' by default, i.e not visible outside
the libac DSO. See this[0] for more details.

We wrap the header file function prototypes in #pragma GCC visibility
push(default)/#pragma GCC visibility pop to make the functions available
outside the library. This is less intrusive than using
__attribute__ ((visibility("default"))) on each prototype individually.

This seems to be recommended for libraries and should help produce more
optimized code.

[0] - https://gcc.gnu.org/wiki/Visibility

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>