ac_fs: Fix ac_fs_is_posix_name()
Commit 4d2c3cf ("ac_fs: Slight code cleanup in ac_fs_is_posix_name()")
made a bit of a meal of things.

We can't increment *name in the while loop, as that will move us past
the first character of the string which may or may not be valid, but it
will also cause us to check the nul character in the switch statement.
We could check for '\0' in the while loop, but that wouldn't fix the
first problem, so do the increment in the switch statement and we won't
see the nul byte.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>