test: Use the __always_unused __attribute__ alias
In commit 41aae54 ("Compile with -Wextra") we changed the test program
compiler options to include -Wno-unused-parameter. However since commit
50de74f ("libac.h: Add some aliases for __attribute__((unused))") we
have some nice shorthand aliases we can use instead of the more verbose
__attribute__((unused)).

So lets make use of the __always_unused alias to mark the various
function arguments that are never used, these are generally callback
functions where we are passed a number of arguments and we may not be
interested in them all. Then we remove the -Wno-unused-parameter option
from the test program compiler command.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>