Build fix's for older versions of glibc & GCC
We were using the GCC option '-fstack-protector-strong' however this is
only available in GCC >= 4.9

When using the clock_*(2) functions you need to link with -lrt for glibc
versions before 2.17

Handle these issues in the Makefile selecting the right options based on
the detected glibc & GCC versions.

Secondly, compiling test.c with GCC 4.4.7 produced the following
warnings

    test.c: In function ‘geo_test’:
    test.c:211: warning: missing initializer
    test.c:211: warning: (near initialization for ‘from.lat’)
    test.c:212: warning: missing initializer
    test.c:212: warning: (near initialization for ‘to.lat’)

it doesn't seem to like initialisers such as

    ac_geo_t from = { 0 };

so switch to using memset(3).

These problems showed up when building under CentOS 6.9, with glibc 2.12
& GCC 4.4.7

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>