ac_str: Add a string split function
Add a function to split a given string on the given delimiter

    char **ac_str_split(const char *string, int delim)

This returns a NULL terminated vector of the split up string. There is
also

    void ac_str_freev(char **stringv)

to free the returned vector.

Empty strings and strings that don't contain any delimiters are returned
as empty vectors, i.e vec[0] == NULL. They should still be free'd with
ac_str_freev.

See the str_test() function in test.c for example usage.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>