ac_str: Add a #define for AC_STR_SPLIT_ALWAYS
Commit 801159c ("ac_str: Change the semantics of ac_str_split()") added
a flags field to ac_str_split(), which could be either 0 or
AC_STR_SPLIT_STRICT, 0 meaning return a field even if there is no
separator character.

Create a #define for this value called AC_STR_SPLIT_ALWAYS to better
document what the intention is. These are really the only two possible
values for this parameter and are used individually so having a value of
0 for one of them shouldn't pose any issues. Yes, it's not really a
flags field...

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>