ac_cqueue: Allow the queue to dynamically grow in size
When adding an item to the queue, ac_cqueue_push() will automatically
grow the queue by AC_CQUEUE_ALLOC_CHUNK_SZ (currently set to 16) if
ac_cqueue_new() was given 0 as the size argument and the queue is
currently full.

Currently we don't try to shrink the queue, this is a little more
involved as the items may not be at the start of the queue's memory
region.

See cqueue_test() in test.c for an example of using this.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>