ac_cqueue: Allow the circular queue to overwrite when full
Up until now the circular queue had two operating modes.

1) A fixed size and when it was full no more items could be added until
   an item was removed.
2) Dynamically grow in size as more items are added.

This commit adds a third option, a fixed size with the overwriting of
the oldest entries.

To this end a flags argument has been added to ac_cqueue_new(), if
passed in as 0, then one of the first two options is used depending on
the size argument.

If the flags argument is passed in as AC_CQUEUE_OVERWRITE then we get
the new behaviour. e.g something like

    queue = ac_cqueue_new(1024, free_queue_item, AC_CQUEUE_OVERWRITE);

This creates a queue that can hold 1024 items and when it is full if
more items are added they will overwrite the oldest items.

See cqueue_test() from test.c for example usage.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>