ac_slist: Add ac_slist_find_custom()
There is a function ac_slist_find() that will find an element in the
list that matches the passed in data, this simply does a pointer
comparison.

It may be useful to have a similar function that doesn't require you to
have a pointer for one the elements in the list to check.

This new function adds an extra argument, which is a pointer to a
comparison function to do the checking.

This comparison function takes two const void * arguments pointing to
the items to check and should return 0 for a match and non-zero for no
match.

The new function looks like

    ac_slist_t *ac_slist_find_custom(ac_slist_t *list,
                                     const void *data,
                                     int (*compar)(const void *a,
                                                   const void *b))

If the item is found it returns a pointer to the list element that
contains the data, NULL otherwise.

You can see an example of this in slist_test() in test.c

We make a slight change to the slist_cmp() function in test.c in that we
alter it so that it is working on pointers to void *, rather than
pointers to pointers of void * which is what we need for qsort(3) for
the list sorting example.

We then create a simple wrapper slist_qcmp() around slist_cmp() to do
the dereferencing of the arguments for passing them into slist_cmp and
use slist_qcmp as the comparison function to qsort.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>