ac_slist: Add ac_slist_remove_custom()
There is a function ac_slist_remove() that will remove an element from
the list that matches the passed in data, this simply does a pointer
comparison.

It may be useful to have a similar function that doesn't require you to
have a pointer for one of the elements in the list to check.

This new function adds an extra argument, which is a pointer to a
comparison function to do the checking.

This comparison function takes two const void * arguments pointing to
the items to check and should return 0 for a match and non-zero for no
match.

The new function looks like

    bool ac_slist_remove_custom(ac_slist_t **list, void *data,
                                int (*compar)(const void *a, const void *b),
                                void (*free_data)(void *data))

You can see an example of this in slist_test() in test.c

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>