ac_htable: Add some Hash Table functions
This adds a simple hash table implementation.

This is based around a new type

    ac_htable_t

and has a similar API to that of GHashTable from glib.

A new hash table is created with the ac_htable_new() function, e.g

    ac_htable_t *htable;

    htable = ac_htable_new(ac_hash_func_str, ac_cmp_str, NULL, NULL);

This creates a hash table that uses ac_hash_func_str() for the hash
function, ac_cmp_str() as the key compare function and the key free
func and data free func set as NULL, i.e they don't need to be free'd.

For the hash function there are three provided functions

    ac_hash_func_str() - create a hash value for a given string
    ac_hash_func_u32() - create a hash value for a given u32
    ac_hash_func_ptr() - create a hash for a given pointer

or you can provide your own hash function that takes the key as a const
void * argument and returns a u32 value.

For the key compare function there also three provided functions

    ac_cmp_str() - compare two strings
    ac_cmp_u32() - compare two u32's
    ac_cmp_ptr() - compare two pointers

Again, you can provide your own key compare function, they take two
const void * arguments pointing to the two keys to compare and should
return a 0 for a match and non-zero otherwise.

The key and data free functions take a void * argument, e.g free(3) is
a suitable function to pass in.

The other hash table functions provided are

    void ac_htable_insert(ac_htable_t *htable, void *key, void *data)
    bool ac_htable_remove(ac_htable_t *htable, void *key)
    void *ac_htable_lookup(const ac_htable_t *htable, const void *key)
    void ac_htable_foreach(ac_htable_t *htable,
                           void (*action)(void *key, void *value,
                                          void *user_data), void *user_data)
    void ac_htable_destroy(ac_htable_t *htable)

With ac_htable_insert(), if you insert items with the same key, the old
item will be removed first (and free'd if required) before inserting the
new item.

You can also get the current number of items in the hash table via
htable->count

This also adds a couple of macros

    AC_LONG_TO_PTR(x)
    AC_PTR_TO_LONG(p)

to convert from a long into a pointer and vice versa. This can be used
to have numbers as your key/data values. e.g

    ac_htable_t *htable;

    htable = ac_htable_new(ac_hash_func_ptr, ac_cmp_ptr, NULL, free);
    ac_htable_insert(htable, AC_LONG_TO_PTR(1), strdup("one"));
    ...

See htable_test() in test.c for example usage.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>