test: net: Add a couple more is in network tests
Add a couple of tests for whether an address falls within the given
network range to cover single host networks.

There may be cases where given something like
2001:db8::dead:beef:cafe/128 and 2001:db8::dead:beef:cafe you want to
check if they are the same.

Yes you could strip the prefix and do a strcmp() but maybe you are
already doing a bunch of such checks and don't want to special case
single hosts.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>