ac_json: Add some functions to help writing JSON
This adds a new type

    ac_jsonw_t

and adds the following functions

    ac_jsonw_init() - initialise a new ac_jsonw_t object

e,g

    ac_jsonw_t *json = ac_jsonw_init();

We can then add various items to the JSON with

    ac_json_add_str()
    ac_json_add_int()
    ac_json_add_float()
    ac_json_add_bool()
    ac_json_add_null()

They should be self explanatory, they all take three arguments, the
ac_jsonw_t to operate on, the name of the field being added or NULL for
when adding array elements and a value.

    ac_json_add_str_or_null()

This function will either insert the given string or null if the value
is NULL or 0 length.

    ac_json_add_array()
    ac_json_add_object()

These functions let you add an array or JSON object.

    ac_json_end_array()
    ac_json_end_object()

These functions close the above off. They can be nested.

    ac_json_end()

This finishes off the JSON.

    ac_json_len()
    ac_json_get()

These are used to retrieve the length of the created JSON and the JSON
itself respectively.

    ac_json_free()

This free's the ac_jsonw_t object.

See json_test() in test.c for example usage.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>