ac_time: Add ac_time_nsleep()
This is a simple wrapper around nanosleep(2) that sleeps through
interrupts (EINTR).

It takes the number of nanoseconds to sleep as its argument.

On error (except for EINTR) it will return -1 and errno can be checked.

There are three defines available to help

    AC_TIME_NS_SEC
    AC_TIME_NS_MSEC
    AC_TIME_NS_USEC

which define the number of nanoseconds in a second, millisecond and
microsecond respectively.

For example usage see time_test() in test.c

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>