ac_json: Allow nameless objects
When trying to create an array of objects using

    ac_json_add_object(json, NULL);

This was generating something like the following

    {
        "object_array": [
            "(null)": {
                    "item1": "Hello",
                    "item2": "World"
            },
            "(null)": {
                    "item1": "qwerty",
                    "item2": "azerty"
            }
        ]
    }

Notice the "(null)"'s. Fix ac_json_add_object() to handle NULL as the
name properly, so that it simply puts "{". Now we get

    {
        "object_array": [
            {
                    "item1": "Hello",
                    "item2": "World"
            },
            {
                    "item1": "qwerty",
                    "item2": "azerty"
            }
        ]
    }

which passes as valid JSON.

Also add a test to json_test() in test.c for this use case.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>