test/time: Fix the ac_time_tspec_diff() test
This test goes through an array of struct timespecs with start and end
times and works out the time difference.

The last items in the array are all 0's to tell the for () loop when to
stop. However we were checking on start.tv_nsec != 0 and one of our test
cases satisfied this so we stopped looping prematurely. What I actually
meant to check on was start.tv_sec and allow all the test cases to run.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>