ac_json: Re-work make_escaped_string()
Rather than operating on the const char *str pointer directly, assign it
to an unsigned char in the while loop. This will allow to handle 8-bit
ascii.

We also rename the add_escaped_char() function to add_escaped_str() as
that is what it takes and add a new add_char() function that operates
similarly but takes an unsigned char and uses memset().

This also allows us to remove the goto from the loop.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>