Makefile: Tighten up compiler and linker flags
This adds -z defs & -z relro to the linker flags

  -z defs to detect under linking
  -z relro for read-only segments after relocation

and

  -Wdeclaration-after-statement, -Wvla & -fexceptions to the compiler
flags

The first two to enforce coding style and -fexceptions to help if libac
is used in multi-threaded programs.

Adding -Wdeclaration-after-statement did cause some issue with some of
the functions in ac_geo.c where variables are being declared all through
the code. In these cases that is perhaps a legitimate thing to do as
they are just implementing a set of equations.

Rather than making a lot of code changes (perhaps not for the best as
stated above) to satisfy this warning we can disable this warning for
specific parts of the code by wrapping it in

  #pragma GCC diagnostic push
  #pragma GCC diagnostic ignored "-Wdeclaration-after-statement"
  ...
  #pragma GCC diagnostic pop

to temporarily disable the warning then re-enable it.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>