ac_queue: Massively re-work ac_cqueue
This was a queue of fixed or dynamic size that was implemented like a
circular buffer. It could operate in a mode whereby it would overwrite
old entries in the queue.

On closer inspection, none of this really worked.

At its heart it was meant to just be a FIFO queue.

So, rename it from cqueue to just queue, remove the fixed or dynamically
sized options (it just grows and shrinks as items are pushed/popped) and
remove the possibility of overwriting elements.

The queue is now implemented on top of the ac_slist_t singly linked list
stuff.

This has resulted in a big simplification of the code. All the functions
have largely remained the same (other than changing from ac_cqueue_* to
ac_queue_*) except that ac_queue_new() now takes no arguments and
ac_queue_destroy() now takes two, the second being the function to free
the elements or NULL for none.

There is also now no ac_queue_is_empty(), just use ac_queue_nr_items().

One optimisation we make when pushing elements into the queue is when
we add them to the tail of queue we track the tail pointer so we don't
need to find the end of the list (like ac_slist_add() would need to do)
each time we want to push something into the queue.

Perhaps something like a circular buffer will be added later.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>