ac_circ_buf: Add a circular buffer
This adds an implementation of a circular buffer based on one in the
Linux Kernel[0][1]

This adds new functions

    ac_circ_buf_new
    ac_circ_buf_count
    ac_circ_buf_pushm
    ac_circ_buf_push
    ac_circ_buf_popm
    ac_circ_buf_pop
    ac_circ_buf_foreach
    ac_circ_buf_reset
    ac_circ_buf_destroy

ac_circ_buf_new creates a new buffer and takes a single size argument,
which must be a power of two. This is to facilitate the use of the
bitwise AND operator for working with the buffer, in an efficient
manner (as opposed to the modulo (%) operator).

Also, the actual usable size will be size-1, as one slot is never used,
this helps in determining when the buffer is full vs empty.

Once the buffer is full, it will reject new items rather than
overwriting old entries.

ac_circ_buf_destroy() will free the memory allocated for the underlying
buffer. It is up to the user to free the memory of items placed into the
buffer. ac_circ_buf_foreach() could be used for that if all the items
that have allocated memory are currently in the buffer.

See the circ_buf_test() function in test.c for example usage.

[0]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/include/linux/circ_buf.h
[0]: https://www.kernel.org/doc/html/latest/core-api/circular-buffers.html

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>