Makefile: Create library symlinks
When linking the library also create a couple of symlinks that are need
when building against it. These are creates when the library is
installed via rpm but can also be useful have in the src/ directory when
doing development/testing against it.

Also adjust the 'clean' target to get rid of all the libac.so entires
and not just the symlinks, this avoids dangling symlinks when the
library version changes.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>