ac_circ_buf: Modify to allow the storing of copied data
The previous interface allowed the storing of pointers (void *) to the
actual data.

However it may be useful to just store the actual data, so you don't
need to worry about the allocating/freeing of dynamic memory which is
likely what those pointers are pointing to.

To do this the ac_circ_buf_new() function has been changed (yes,
unfortunately this is an API break) to take a second argument, the
element size, this is the size in bytes of the individual items that
will be stored in the buffer. Set this to 0 to get the previous
behaviour for storing pointers.

See circ_buf_test() in test.c for some example usage.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>