ac_misc: Add a shuffle function
Add an ac_misc_shuffle() function which is just a thin wrapper around
any number of functions implementing various shuffle algorithms.

Currently there is only an implementation of the Fisher-Yates[0]
algorithm based on code from[1].

It takes arguments similar to qsort(3), except for the compar function.
It looks like

    ac_misc_shuffle(void *base, size_t nmemb, size_t size,
                    ac_misc_shuffle_t algo)

The first three arguments are the same as for qsort(); the array to
shuffle, the number of elements in the array and the size of each
element.

algo specifies the shuffle algorithm to use, currently only
AC_MISC_SHUFFLE_FISHER_YATES is supported.

See the misc_test() function in test.c for example usage.

[0]: https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle
[1]: https://phoxis.org/2015/04/21/a-generic-fisher-yates-shuffle/

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>