ac_misc: Use random_r(3) instead of random(3)
The Linux random(3) man page is somewhat misleading, in that it seems to
suggest random() is thread-safe under 'ATTRIBUTES'. Then down in the
'NOTES' section it says

    The  random()  function  should  not  be used in multithreaded programs
    where reproducible behavior is required.  Use random_r(3) for that pur‐
    pose

So, OK, lets use random_r(3) instead. This involved the use of
initstate_r(3) and *we* then don't require the use of srandom_r(3).

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>