ac_btree: Fix dangling pointer return from ac_btree_remove()
When running some code that uses this through valgrind I was seeing the
following

    ...
    ==54709== Invalid read of size 8
    ==54709==    at 0x4879BA4: ac_btree_remove (ac_btree.c:153)
    ==54709==    by 0x4011A0: main (bt.c:79)
    ==54709==  Address 0x4bc7580 is 0 bytes inside a block of size 24 free'd
    ==54709==    at 0x4839A0C: free (vg_replace_malloc.c:540)
    ==54709==    by 0x4976CC2: tdelete (tsearch.c:676)
    ==54709==    by 0x4879B95: ac_btree_remove (ac_btree.c:148)
    ==54709==    by 0x4011A0: main (bt.c:79)
    ==54709==  Block was alloc'd at
    ==54709==    at 0x483880B: malloc (vg_replace_malloc.c:309)
    ==54709==    by 0x49767F9: tsearch (tsearch.c:338)
    ==54709==    by 0x4879B65: ac_btree_add (ac_btree.c:126)
    ==54709==    by 0x401148: main (bt.c:61)
    ...

This occurred after removing the root node (i.e there was only one node
in the tree), as the tsearch(3) man page says

    tdelete()  returns a pointer to the parent of the node deleted, or NULL
    if the item was not found.  If the deleted  node  was  the  root  node,
    tdelete() returns a dangling pointer that must not be accessed.

and we were dereferencing that pointer when returning it.

The simple fix here is to just return NULL in such cases. However that
does then have the problem that you won't immediately know if you got
NULL because the item to be removed couldn't be found or if it was the
last item removed.

Still that is better than giving the user some pointer to random memory
that they might try and use.

A subsequent commit will add an ac_btree_is_empty() function that can be
used to help inform of such cases.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>