ac_quark: Make it actually work
This only seemed to work under basic testing. However in actual use it
turned out to be somewhat broken. It would create new quarks fine, but
when trying to look up a string given an id, it would find some quarks
but not others.

This was due to the broken use of ac_btree_* and trying to use different
indexes to store the quarks (strings) and to lookup the quarks by id
(integers), this just doesn't work. I guess due to the fact that the
tree created using strings does not match the tree trying to be
traversed using id's.

What we need to do is also keep another mapping of id's to quarks, this
is done just as a dynamically resized array.

So when creating a new quark we store the quark in the btree and also in
the array at the index of its id. Of course what's actually being stored
is just pointers to a quark_node structure.

Then when looking up a quark given an id, we simply pull it out of the
array.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>