ac_json: Add a couple of functions for loading JSON from a file
This adds

    char *ac_json_load_from_fd(int fd, off_t offset)
    char *ac_json_load_from_file(const char *file, off_t offset)

for loading both from a given file path and also by an already open file
descriptor.

Both functions return a pointer to a memory buffer containing the JSON
which should be free'd when done with.

Also both functions take on offset parameter to allow the loading of
JSON from arbitrary locations within the file. Pass 0 for the start.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>