ac_btree: Adjust the ac_btree_foreach() interface
This is a function to iterate over the given tree applying the specified
action function on each node.

Underneath this uses the twalk(3) function which has the following
prototype

    void twalk(const void *root,
               void (*action)(const void *nodep, VISIT which,
                              int depth));

which doesn't allow to pass user data to the action() function, which
can often be useful without resorting to global variables.

However there is a twalk_r(3) function with the following prototype

    void twalk_r(const void *root,
                 void (*action)(const void *nodep, VISIT which,
                                void *closure),
                 void *closure);

which essentially replaces the depth argument with a user data one.

So we loose the depth argument to our action function but can now pass
user data into it if required which I think will be a lot more useful.
If we ever find we want/need the depth argument then we can perhaps do
something like this[0].

We also adjust some of the function parameters to match the above. i.e
mark the tree root as const and remove const from the VISIT which
argument, this is just a typedef to an enum.

We also update test.c to show the use of the 'which' variable in the
action() function.

Also make it clear that this is a 'binary search tree'.

[0]: https://www.gnu.org/software/libc/manual/html_node/Tree-Search-Function.html#index-twalk_005fr

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>