Makefile: Re-work the use of the address sanitizer
The previous commit added the ability to build with the address
sanitizer, libasan.

However, only the executable really needs to be compiled/linked with it.
We also don't need to explicitly link with -lasan as this will get added
automatically. Indeed it is better this way as it seems libasan needs to
be the first library loaded and this will ensure that.

So rather than put -fsanitize=address in CFLAGS which will be used by
*all* the source files. Create a new ASAN variable for this purpose and
use it only in the test binary recipe.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>