ac_*: Make various function arguments const
In a number of the API functions we can mark the main object passed in
as const (quite a few already did this).

This is true in cases where we are not modifying the thing being pointed
to.

Perhaps surprisingly this includes some of the _destroy/_free functions.

Some of them are really just a wrapper around free(3) and it is
regarded (by some at least[0][1], and I would agree) that the argument
to free() should have been 'const void *' rather than just 'void *'.

This is largely because free() doesn't actually change the data pointed
to. And as Linus says[0]

    "Anything that *can* take a const pointer should always do so."

[0]: https://yarchive.net/comp/const.html
[1]: https://nullprogram.com/blog/2015/02/17/

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>